pyproject.toml
[project]
name = "portfolio-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apimoex>=1.4.0",
    "loguru>=0.7.3",
    "moexalgo>=2.3.2",
    "pandas>=2.3.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
    "yfinance>=0.2.65",
]
